requirements.txt
tensorflow==2.15
streamlit
streamlit-drawable-canvas
numpy
pandas
pillow
scikit-learn
tensorflow-cpu
